+++ packages/moonstone/ExpandableItem/ExpandableItem.js
@@ -14,6 +14,7 @@ import kind from '@enact/core/kind';
 import React from 'react';
 import PropTypes from 'prop-types';
 import Spotlight from '@enact/spotlight';
+import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
 
 import LabeledItem from '../LabeledItem';
 
@@ -22,6 +23,8 @@ import ExpandableTransitionContainer from './ExpandableTransitionContainer';
 
 const isUp = is('up');
 const isDown = is('down');
+
+const ContainerDiv = SpotlightContainerDecorator('div');
 
 /**
  * {@link moonstone/ExpandableItem.ExpandableItemBase} is a stateless component that
@@ -320,12 +323,12 @@ const ExpandableItemBase = kind({
 		const ariaProps = extractAriaProps(rest);
 
 		return (
-			<div
+			<ContainerDiv
 				{...rest}
 				aria-disabled={disabled}
 				disabled={disabled}
 				open={open}
-				ref={setContainerNode}
+				spotlightContainerRef={setContainerNode}
 			>
 				<LabeledItem
 					{...ariaProps}
@@ -353,7 +356,7 @@ const ExpandableItemBase = kind({
 				>
 					{children}
 				</ExpandableTransitionContainer>
-			</div>
+			</ContainerDiv>
 		);
 	}
 });