+++ packages/spotlight/SpotlightContainerDecorator/SpotlightContainerDecorator.js
@@ -118,6 +118,14 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			containerId: PropTypes.string,
 
 			/**
+			 * Function to capture a reference to the spotlight container node.
+			 *
+			 * @type {Function}
+			 * @public
+			 */
+			spotlightContainerRef: PropTypes.func,
+
+			/**
 			 * When `true`, controls in the container cannot be navigated.
 			 *
 			 * @type {Boolean}
@@ -232,7 +240,7 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		render () {
-			const {spotlightDisabled, spotlightMuted, ...rest} = this.props;
+			const {spotlightContainerRef, spotlightDisabled, spotlightMuted, ...rest} = this.props;
 			delete rest.containerId;
 			delete rest.spotlightRestrict;
 
@@ -248,7 +256,7 @@ const SpotlightContainerDecorator = hoc(defaultConfig, (config, Wrapped) => {
 				rest['data-container-muted'] = spotlightMuted;
 			}
 
-			return <Wrapped {...rest} />;
+			return <Wrapped {...rest} ref={spotlightContainerRef} />;
 		}
 	};
 });